Split status markdown smoke fixtures
