+++ examples/control_plane/status_markdown_fixtures.py
@@ -5,6 +5,8 @@ from __future__ import annotations
 import json
 from pathlib import Path
 
+from loopx.status import render_status_markdown
+
 
 OLD_PLANNED_ACTION = "先审阅 LoopX operator gate；同意后再发送项目 agent 命令"
 NEW_PLANNED_ACTION = "先在 LoopX 完成 operator 判断；同意后项目 Agent 只执行 read-only map dry-run"
@@ -32,6 +34,247 @@ DEPENDENCY_BLOCKER_GOAL_ID = "dependency-owner-gate"
 DEPENDENCY_AGENT_TODO = "Continue the gate-independent P1/P2 product-hardening slice."
 DEPENDENCY_MONITOR_TODO = "Side-bypass dependency monitor: observe public-safe replay transitions only."
 DEPENDENCY_USER_TODO = "Confirm the sibling project evidence gate before its controller resumes delivery."
+
+
+def assert_missing_project_asset_markdown_fallback() -> None:
+    markdown = render_status_markdown({
+        "ok": True,
+        "registry": "./fixtures/registry.json",
+        "runtime_root": "./fixtures/runtime",
+        "goal_count": 1,
+        "run_count": 0,
+        "contract": {"ok": True, "summary": {"errors": 0, "warnings": 0, "checks": 0}},
+        "global_registry": {"available": False, "summary": {}},
+        "attention_queue": {
+            "item_count": 1,
+            "needs_user_or_controller": 0,
+            "needs_controller": 0,
+            "needs_codex": 1,
+            "watching_external_evidence": 0,
+            "items": [
+                {
+                    "goal_id": "legacy-status-only",
+                    "status": "state_refreshed",
+                    "lifecycle_phase": "refreshed",
+                    "waiting_on": "codex",
+                    "severity": "action",
+                    "source": "latest_run",
+                    "recommended_action": "raw queue action should be labeled as fallback",
+                }
+            ],
+        },
+        "run_history": {"goals": []},
+    })
+
+    assert "project_asset_source: legacy/raw fallback" in markdown, markdown
+    assert "owner/gate/stop are not project_asset-backed" in markdown, markdown
+    assert "project_asset: owner=" not in markdown, markdown
+
+
+def assert_usage_summary_markdown() -> None:
+    markdown = render_status_markdown({
+        "ok": True,
+        "registry": "./fixtures/registry.json",
+        "runtime_root": "./fixtures/runtime",
+        "goal_count": 2,
+        "run_count": 4,
+        "contract": {"ok": True, "summary": {"errors": 0, "warnings": 0, "checks": 0}},
+        "global_registry": {"available": False, "summary": {}},
+        "usage_summary": {
+            "available": True,
+            "source": "run_history",
+            "sample_run_count": 4,
+            "totals": {
+                "runs_24h": 4,
+                "runs_7d": 4,
+                "quota_spend_slots_24h": 2,
+                "quota_spend_slots_7d": 2,
+                "automation_run_count_24h": 2,
+                "automation_run_count_7d": 2,
+                "progress_signal_run_count_24h": 1,
+                "progress_signal_run_count_7d": 1,
+            },
+            "goals": [
+                {
+                    "goal_id": "meta-loop",
+                    "runs_24h": 3,
+                    "runs_7d": 3,
+                    "quota_spend_slots_24h": 2,
+                    "automation_run_count_24h": 2,
+                    "progress_signal_run_count_24h": 0,
+                    "project_share_24h": 0.75,
+                },
+                {
+                    "goal_id": "delivery-signal",
+                    "runs_24h": 1,
+                    "runs_7d": 1,
+                    "quota_spend_slots_24h": 0,
+                    "automation_run_count_24h": 0,
+                    "progress_signal_run_count_24h": 1,
+                    "project_share_24h": 0.25,
+                },
+            ],
+        },
+        "attention_queue": {
+            "item_count": 0,
+            "needs_user_or_controller": 0,
+            "needs_controller": 0,
+            "needs_codex": 0,
+            "watching_external_evidence": 0,
+            "items": [],
+        },
+        "run_history": {"goals": []},
+    })
+
+    assert "## Usage Summary" in markdown, markdown
+    assert "samples=4" in markdown, markdown
+    assert "quota_slots_24h=2" in markdown, markdown
+    assert "progress_signals_24h=1" in markdown, markdown
+    assert "`meta-loop`: runs_24h=3" in markdown, markdown
+    assert "`delivery-signal`: runs_24h=1" in markdown, markdown
+    assert "progress_signals_24h=0" in markdown, markdown
+    assert "progress_signals_24h=1" in markdown, markdown
+
+
+def assert_event_ledger_summary_markdown() -> None:
+    markdown = render_status_markdown({
+        "ok": True,
+        "registry": "./fixtures/registry.json",
+        "runtime_root": "./fixtures/runtime",
+        "goal_count": 2,
+        "run_count": 5,
+        "contract": {"ok": True, "summary": {"errors": 0, "warnings": 0, "checks": 0}},
+        "global_registry": {"available": False, "summary": {}},
+        "event_ledger_summary": {
+            "available": True,
+            "source": "run_history",
+            "sample_run_count": 5,
+            "event_classes": ["accounting", "decision", "evidence", "state", "work"],
+            "totals": {
+                "events_24h": 5,
+                "events_7d": 5,
+                "by_class_24h": {
+                    "accounting": 1,
+                    "decision": 1,
+                    "evidence": 1,
+                    "state": 1,
+                    "work": 1,
+                },
+                "by_class_7d": {
+                    "accounting": 1,
+                    "decision": 1,
+                    "evidence": 1,
+                    "state": 1,
+                    "work": 1,
+                },
+            },
+            "goals": [
+                {
+                    "goal_id": "meta-loop",
+                    "events_24h": 3,
+                    "events_7d": 3,
+                    "latest_event_class": "work",
+                    "by_class_24h": {
+                        "accounting": 1,
+                        "decision": 0,
+                        "evidence": 0,
+                        "state": 1,
+                        "work": 1,
+                    },
+                },
+                {
+                    "goal_id": "operator-gate",
+                    "events_24h": 2,
+                    "events_7d": 2,
+                    "latest_event_class": "decision",
+                    "by_class_24h": {
+                        "accounting": 0,
+                        "decision": 1,
+                        "evidence": 1,
+                        "state": 0,
+                        "work": 0,
+                    },
+                },
+            ],
+        },
+        "attention_queue": {
+            "item_count": 0,
+            "needs_user_or_controller": 0,
+            "needs_controller": 0,
+            "needs_codex": 0,
+            "watching_external_evidence": 0,
+            "items": [],
+        },
+        "run_history": {"goals": []},
+    })
+
+    assert "## Event Ledger Summary" in markdown, markdown
+    assert "samples=5" in markdown, markdown
+    assert "events_24h=5" in markdown, markdown
+    assert "classes_24h=accounting=1 decision=1 evidence=1 state=1 work=1" in markdown, markdown
+    assert "`meta-loop`: events_24h=3" in markdown, markdown
+    assert "latest=work" in markdown, markdown
+    assert "`operator-gate`: events_24h=2" in markdown, markdown
+    assert "latest=decision" in markdown, markdown
+
+
+def assert_decision_freshness_summary_markdown() -> None:
+    markdown = render_status_markdown({
+        "ok": True,
+        "registry": "./fixtures/registry.json",
+        "runtime_root": "./fixtures/runtime",
+        "goal_count": 2,
+        "run_count": 5,
+        "contract": {"ok": True, "summary": {"errors": 0, "warnings": 0, "checks": 0}},
+        "global_registry": {"available": False, "summary": {}},
+        "decision_freshness_summary": {
+            "available": True,
+            "source": "run_history",
+            "sample_run_count": 5,
+            "window_days": 7,
+            "summary": {
+                "decision_count": 2,
+                "stale_count": 1,
+                "rebase_required_count": 2,
+                "fresh_count": 0,
+            },
+            "items": [
+                {
+                    "goal_id": "meta-loop",
+                    "decision_kind": "human_reward",
+                    "decision_at": "2026-05-20T12:00:00+00:00",
+                    "age_days": 8.25,
+                    "newer_event_count_7d": 3,
+                    "freshness_state": "stale_rebase_required",
+                },
+                {
+                    "goal_id": "operator-gate",
+                    "decision_kind": "operator_gate",
+                    "decision_at": "2026-05-27T12:00:00+00:00",
+                    "age_days": 0.25,
+                    "newer_event_count_7d": 1,
+                    "freshness_state": "rebase_required",
+                },
+            ],
+        },
+        "attention_queue": {
+            "item_count": 0,
+            "needs_user_or_controller": 0,
+            "needs_controller": 0,
+            "needs_codex": 0,
+            "watching_external_evidence": 0,
+            "items": [],
+        },
+        "run_history": {"goals": []},
+    })
+
+    assert "## Decision Freshness Summary" in markdown, markdown
+    assert "window_days=7" in markdown, markdown
+    assert "decisions=2" in markdown, markdown
+    assert "stale=1" in markdown, markdown
+    assert "rebase_required=2" in markdown, markdown
+    assert "`meta-loop`: kind=human_reward state=stale_rebase_required" in markdown, markdown
+    assert "`operator-gate`: kind=operator_gate state=rebase_required" in markdown, markdown
 
 
 def write_planned_registry(root: Path) -> Path:
